+++ semialign/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Split `Semialign` into `Semialign` and `Zip`.
 - Rename old `Zip` into `Repeat`
 - i.e. current main hierarchy is
+- Remove `malign`, use `salign` or `alignWith mappend` where `Monoid` is necessary.
 
 ```haskell
 instance Functor f => Semialign f where